migrate analytics params from clientId/accountId to teamId/projectId
