+++ apps/dashboard/src/app/team/[team_slug]/(team)/~/analytics/page.tsx
@@ -5,21 +5,18 @@ import {
   getWalletConnections,
   getWalletUsers,
 } from "@/api/analytics";
-import { redirect } from "next/navigation";
-
-import type {
-  InAppWalletStats,
-  WalletStats,
-  WalletUserStats,
-} from "types/analytics";
-
 import {
   type DurationId,
   type Range,
   getLastNDaysRange,
 } from "components/analytics/date-range-selector";
-
+import { redirect } from "next/navigation";
 import { type WalletId, getWalletInfo } from "thirdweb/wallets";
+import type {
+  InAppWalletStats,
+  WalletStats,
+  WalletUserStats,
+} from "types/analytics";
 import { AnalyticsHeader } from "../../../../components/Analytics/AnalyticsHeader";
 import { CombinedBarChartCard } from "../../../../components/Analytics/CombinedBarChartCard";
 import { EmptyState } from "../../../../components/Analytics/EmptyState";
@@ -27,8 +24,6 @@ import { PieChartCard } from "../../../../components/Analytics/PieChartCard";
 
 import { getTeamBySlug } from "@/api/team";
 import { GenericLoadingPage } from "@/components/blocks/skeletons/GenericLoadingPage";
-import type { Account } from "@3rdweb-sdk/react/hooks/useApi";
-import { getValidAccount } from "app/account/settings/getAccount";
 import { EmptyStateCard } from "app/team/components/Analytics/EmptyStateCard";
 import { Suspense } from "react";
 import { TotalSponsoredChartCardUI } from "../../_components/TotalSponsoredCard";
@@ -51,7 +46,6 @@ export default async function TeamOverviewPage(props: {
     props.searchParams,
   ]);
 
-  const account = await getValidAccount(`/team/${params.team_slug}`);
   const team = await getTeamBySlug(params.team_slug);
 
   if (!team) {
@@ -75,7 +69,6 @@ export default async function TeamOverviewPage(props: {
         <Suspense fallback={<GenericLoadingPage />}>
           <OverviewPageContent
             teamId={team.id}
-            account={account}
             range={range}
             interval={interval}
             searchParams={searchParams}
@@ -88,12 +81,11 @@ export default async function TeamOverviewPage(props: {
 
 async function OverviewPageContent(props: {
   teamId: string;
-  account: Account;
   range: Range;
   interval: "day" | "week";
   searchParams: SearchParams;
 }) {
-  const { teamId, account, range, interval, searchParams } = props;
+  const { teamId, range, interval, searchParams } = props;
 
   const [
     walletConnections,
@@ -113,7 +105,7 @@ async function OverviewPageContent(props: {
     }),
     // Time series data for wallet users
     getWalletUsers({
-      accountId: account.id,
+      teamId: teamId,
       from: range.from,
       to: range.to,
       period: interval,
