+++ apps/dashboard/src/app/team/[team_slug]/[project_slug]/connect/in-app-wallets/layout.tsx
@@ -23,7 +23,7 @@ export default async function Layout(props: {
 
   return (
     <div>
-      <InAppWalletsHeader clientId={project.publishableKey} />
+      <InAppWalletsHeader teamId={project.teamId} projectId={project.id} />
       <div className="h-8" />
 
       <TabPathLinks
